refact: redesign FrameToArrayConverter
Fix test
