+++ .github/workflows/ci.yml
@@ -115,8 +115,8 @@ jobs:
       - name: run pytest
         run: |
           if [ "${{ matrix.os }}" == "ubuntu-latest" ]; then
-            pytest --ignore=araviq6/tests/test_array2qvideoframe.py
-            pytest araviq6/tests/test_array2qvideoframe.py
+            pytest --ignore=araviq6/tests/test_videostream.py
+            pytest araviq6/tests/test_videostream.py
           else
             pytest
           fi
